Package and asset system
Entirely new package and object serialization system, objects are lazy-loaded from their respective packages when the package is loaded and an asset is requested.

+++ Lumina/Engine/Source/Lumina.h
@@ -25,5 +25,5 @@
 
 
 // Invalid Index
-constexpr int8 INDEX_NONE = -1;
+constexpr SIZE_T INDEX_NONE = -1;
